+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Agonising Blast enchantment.
 - Potion of healing versions would not import both in Core Rules for 2014 and 2024 were both enabled.
 - When using 2014 5e rules, weapon property description will be removed from weapon items, e.g. Longsword.
+- Spells with a mention of having a Study action will generate an Investigation activity.
 
 # 6.5.7